+++ src/components/pages/CompanyWallet/RuleCard.tsx
@@ -17,22 +17,26 @@
  * SPDX-License-Identifier: Apache-2.0
  ********************************************************************************/
 
-import { Typography } from '@catena-x/portal-shared-components'
+import { Tooltips, Typography } from '@catena-x/portal-shared-components'
 import './CompanyWallet.scss'
 import {
   CredentialSubjectStatus,
+  CredentialType,
   type WalletContent,
 } from 'features/compayWallet/companyWalletApiSlice'
 import dayjs from 'dayjs'
 import { useTranslation } from 'react-i18next'
 import { Box, Grid } from '@mui/material'
+import SettingsBackupRestoreIcon from '@mui/icons-material/SettingsBackupRestore'
 
 type Hash<T> = Record<string, T>
 
 export default function RuleCard({
   sections,
+  handleRevoke,
 }: {
   readonly sections: Hash<WalletContent[]>
+  readonly handleRevoke: (id: string) => void
 }): JSX.Element {
   const { t } = useTranslation()
   const keys = Object.keys(sections)
@@ -53,6 +57,14 @@ export default function RuleCard({
     status === CredentialSubjectStatus.REVOKED
       ? CredentialSubjectStatus.INACTIVE
       : status
+
+  const canShowRevoke = (item: WalletContent) => {
+    return (
+      item.status === CredentialSubjectStatus.ACTIVE &&
+      item?.credentialType !== CredentialType.MEMBERSHIP &&
+      item.credentialType !== CredentialType.BUSINESS_PARTNER_NUMBER
+    )
+  }
 
   return (
     <>
@@ -78,21 +90,21 @@ export default function RuleCard({
             }}
             className="grid-layout"
           >
-            {sections[key]?.map((item: WalletContent) => (
+            {sections[key]?.map((item: WalletContent, index) => (
               <Grid
                 item
                 xs={4}
                 sm={4}
                 md={4}
                 className="main-rule-card-container"
-                key={item.expiryDate}
+                key={item.expiryDate ?? index}
                 sx={{
                   paddingLeft: '20px !important',
                 }}
               >
                 <Box className="gradient-container">
                   <Box
-                    key={item.expiryDate}
+                    key={item.expiryDate ?? index}
                     className="rule-card-container"
                     sx={{
                       backgroundColor:
@@ -128,15 +140,37 @@ export default function RuleCard({
                     <Typography className="text" variant="h4">
                       {item?.credentialType.split('_').join(' ')}
                     </Typography>
-                    <div>
-                      <Typography className="text" variant="body3">
-                        {item.authority}
-                      </Typography>
-                      <Typography className="text" variant="body3">
-                        {t('content.companyWallet.expiry')}
-                        {dayjs(item.expiryDate).format('YYYY-MM-DD')}
-                      </Typography>
-                    </div>
+                    <Box className="revoke-container">
+                      <div>
+                        <Typography className="text" variant="body3">
+                          {item.authority}
+                        </Typography>
+                        <Typography className="text" variant="body3">
+                          {t('content.companyWallet.expiry')}
+                          {dayjs(item.expiryDate).format('YYYY-MM-DD')}
+                        </Typography>
+                      </div>
+                      {canShowRevoke(item) && (
+                        <Tooltips
+                          tooltipPlacement="top-start"
+                          tooltipText={t('content.companyWallet.revokeHint')}
+                          children={
+                            <Box
+                              onClick={() => {
+                                handleRevoke(item.credentialDetailId)
+                              }}
+                            >
+                              <SettingsBackupRestoreIcon
+                                sx={{
+                                  color: '#c84f4f',
+                                  cursor: 'pointer',
+                                }}
+                              />
+                            </Box>
+                          }
+                        />
+                      )}
+                    </Box>
                   </Box>
                 </Box>
               </Grid>