+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## 2.0.0-RC9
+
+### Feature
+
+- Company Wallet
+  - Revoke credential
+
+### Bugfix
+
+- App Subscription
+  - Blank Page fix on autosetup call
+
 ## 2.0.0-RC8
 
 ### Feature